Model Backdrop was "click", now "mousedown"
As mentioned in #285 when someone starts a click event inside the modal content area, and then releases it outside, in the backdrop, the modal hides. This is a frustrating issue when selecting text in the modal, as I often do. This only happens in chrome and is also an issue in bootstrap jquery itself. 

This is the simplest possible fix - replacing @click.self with @mousedown.self. This avoids the weird chrome behaviour.