Model Backdrop was "click", now "mousedown" (#304)
* Model Backdrop was "click", now "mousedown"

As mentioned in #285 when someone starts a click event inside the modal content area, and then releases it outside, in the backdrop, the modal hides. This is a frustrating issue when selecting text in the modal, as I often do. This only happens in chrome and is also an issue in bootstrap jquery itself. 

This is the simplest possible fix - replacing @click.self with @mousedown.self. This avoids the weird chrome behaviour.

* Changed MessageBox to use mousedown

'should be able to set alert backdrop to true' was sending click event, which unfortunately excludes mousedown. Changed it to send the full event sequence (mousedown,mouseup, click) for maximum generality.

* Modal.spec.js - Mousedown changes.

Changed backdrop click tests to send the full event sequence (mousedown, mouseup, click) rather than just click.